refactor(connector): merge api and domparser packages
