+++ src/main/kotlin/com/github/thetric/iliasdownloader/cli/SyncSettings.kt
@@ -1,7 +1,7 @@
 package com.github.thetric.iliasdownloader.cli
 
 import com.github.thetric.iliasdownloader.cli.preferences.UserPreferences
-import com.github.thetric.iliasdownloader.connector.api.model.Course
+import com.github.thetric.iliasdownloader.connector.model.Course
 
 /**
  * Temporary sync settings. The settings are persisted as [UserPreferences].
