+++ src/main/kotlin/com/github/thetric/iliasdownloader/connector/IliasHttpException.kt
@@ -1,0 +1,9 @@
+package com.github.thetric.iliasdownloader.connector
+
+import com.github.thetric.iliasdownloader.connector.exception.IliasException
+
+class IliasHttpException(
+    override val message: String,
+    val url: String,
+    val statusCode: Int
+) : IliasException(message)
