+++ src/main/kotlin/com/github/thetric/iliasdownloader/connector/model/CourseFolder.kt
@@ -1,4 +1,4 @@
-package com.github.thetric.iliasdownloader.connector.api.model
+package com.github.thetric.iliasdownloader.connector.model
 
 /**
  * Course folder in an Ilias [Course].
