+++ src/main/kotlin/com/github/thetric/iliasdownloader/connector/model/LoginCredentials.kt
@@ -1,9 +1,9 @@
-package com.github.thetric.iliasdownloader.connector.api.model
+package com.github.thetric.iliasdownloader.connector.model
 
 /**
  * Represents the login credentials necessary for an Ilias login.
  *
- * @see com.github.thetric.iliasdownloader.connector.api.IliasService
+ * @see com.github.thetric.iliasdownloader.connector.IliasService
  */
 data class LoginCredentials(
     val userName: String,
